+++ data/index.js
@@ -1,0 +1,24 @@
+import Africa from './Africa.js'
+import America from './America.js'
+import Antarctica from './Antarctica.js'
+import Asia from './Asia.js'
+import Atlantic from './Atlantic.js'
+import Australia from './Australia.js'
+import Etc from './Etc.js'
+import Europe from './Europe.js'
+import Indian from './Indian.js'
+import Pacific from './Pacific.js'
+
+let zones = Object.assign({},
+  Africa,
+  America,
+  Antarctica,
+  Asia,
+  Atlantic,
+  Australia,
+  Etc,
+  Europe,
+  Indian,
+  Pacific,
+)
+export default zones