+++ changelog.md
@@ -1,4 +1,12 @@
-### 1.4.0 [July 2021]
+### 1.5.0 [April 2023]
+- **[fix]** - inverted GMT zones #21
+- **[change]** - add GMT+14, -14
+- **[change]** - update dependencies
+
+### 1.4.1 [August 2022]
+- **[change]** - move ireland off of british standard time
+
+### 1.4.0 [July 2022]
 
 - **[change]** - internal refactor
 - **[change]** - add dst date